trying qtbot, not succesful, refactor
